requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
chromadb==0.4.18
langchain==0.0.350
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.2 